fix: 修复编译错误，成功构建 APK

+++ app/src/main/java/com/duoji/app/ui/components/animation/AnimatedSection.kt
@@ -26,7 +26,7 @@ fun AnimatedSection(
             animationSpec = tween(durationMillis = animDuration, delayMillis = delayMillis)
         ) + slideInVertically(
             animationSpec = tween(durationMillis = animDuration, delayMillis = delayMillis)
-        ) { offsetPx },
+        ) { offsetPx.toInt() },
         modifier = modifier
     ) {
         content()
