+++ app/src/main/java/com/duoji/app/ui/components/animation/StaggeredListItem.kt
@@ -46,7 +46,7 @@ fun StaggeredListItem(
                 durationMillis = animDuration,
                 delayMillis = index * delayPerItem
             )
-        ) { offsetPx },
+        ) { offsetPx.toInt() },
         modifier = modifier
     ) {
         content()
